initial responsive re-org to accomodate outcome trade in outcome row
