--- src/modules/outcomes/less/outcome-row.less
@@ -1,8 +1,0 @@
-.outcome-row {
-
-	.table-row();
-
-	&.link {
-		cursor: default;
-	}
-}
